fix: pretty dom not passing the element to be serialized
